fixing version
