+++ frontend/components/media/Histogram.tsx
