+++ frontend/components/search/AlbumSearchCard.tsx
